Update: 支出・収入報告フォームを共通コンポーネント化
